refactor: restructure to multi-module layout with CQRS commands
  Move all invoice code into modules/invoice/ bounded context. Consolidate
  domain models with their enums and services. Split monolithic use cases
  (ProcessOrderMessage, ProcessInvoiceMessage) into four CQRS commands
  (create, sign, send, authorize) with dedicated message handlers that
  dispatch by invoice status.
