--- src/modules/app/dtos/index.ts
@@ -1,4 +1,0 @@
-export * from './core';
-export * from './invoice';
-export * from './message';
-export * from './sri';
