--- src/modules/app/dtos/invoice.ts
@@ -1,13 +1,0 @@
-export interface InvoiceDTO {
-  id: string;
-  orderId: string;
-  accessCode: string;
-  status: string;
-  signatureId: string;
-  xml: string;
-  statusHistory: {
-    name: string;
-    statusDate: string;
-    description?: string;
-  }[];
-}
