--- src/modules/app/mappers/index.ts
@@ -1,3 +1,0 @@
-export * from './core';
-export * from './invoice';
-export * from './sri';
