--- src/modules/app/usecase.ts
@@ -1,97 +1,0 @@
-import {
-  AuthorizationVoucherStatus,
-  InvoiceStatus,
-  ValidationVoucherStatus,
-} from '../domain/constants';
-import {
-  CorePort,
-  MessageProducer,
-  SealifyPort,
-  SriAuthorizationPort,
-  SriValidationPort,
-} from '../domain/ports';
-import { Invoice, InvoiceMessage, OrderMessage } from '../domain/models';
-
-import { InvoiceRepository } from '../domain/repositories';
-import { logger } from '#/shared/logger';
-
-export interface ProcessMessageUseCase<T> {
-  execute(message: T): Promise<void>;
-}
-
-export class ProcessOrderMessage implements ProcessMessageUseCase<OrderMessage> {
-  constructor(
-    private corePort: CorePort,
-    private invoiceRepository: InvoiceRepository,
-    private messageProducer: MessageProducer<InvoiceMessage>,
-  ) {}
-
-  async execute(message: OrderMessage) {
-    const order = await this.corePort.retrieveOrder(message.id);
-    const xml = order.generateInvoiceXml();
-    const invoice = new Invoice(
-      undefined,
-      order.id.toString(),
-      order.accessCode,
-      InvoiceStatus.CREATED,
-      order.sriConfig.signature,
-      xml,
-    );
-    invoice.addStatusHistory(InvoiceStatus.CREATED, new Date(), 'XML created');
-    await this.invoiceRepository.createInvoice(invoice);
-    await this.messageProducer.sendMessage(
-      new InvoiceMessage(invoice.id, invoice.orderId, invoice.status),
-    );
-  }
-}
-
-export class ProcessInvoiceMessage implements ProcessMessageUseCase<InvoiceMessage> {
-  constructor(
-    private sealifyPort: SealifyPort,
-    private sriValidationPort: SriValidationPort,
-    private sriAuthorizationPort: SriAuthorizationPort,
-    private invoiceRepository: InvoiceRepository,
-    private messageProducer: MessageProducer<InvoiceMessage>,
-  ) {}
-
-  async execute(message: InvoiceMessage) {
-    const invoice = await this.invoiceRepository.getInvoiceById(message.id);
-    if (!invoice) {
-      logger.warn({ invoiceId: message.id }, 'Invoice not found');
-      return;
-    }
-
-    if ([InvoiceStatus.REJECTED, InvoiceStatus.AUTHORIZED].includes(invoice.status)) {
-      logger.warn({ invoiceId: message.id }, 'Invoice already processed');
-      return;
-    }
-
-    if (invoice.status === InvoiceStatus.CREATED) {
-      const signedXml = await this.sealifyPort.sealInvoice(invoice.xml, invoice.signatureId);
-      invoice.xml = signedXml;
-      invoice.addStatusHistory(InvoiceStatus.SIGNED, new Date(), 'XML signed');
-    } else if (invoice.status === InvoiceStatus.SIGNED) {
-      const validationVoucher = await this.sriValidationPort.validateXml(invoice.xml);
-      const invoiceStatus =
-        validationVoucher.status === ValidationVoucherStatus.ACCEPTED
-          ? InvoiceStatus.SENT
-          : InvoiceStatus.REJECTED;
-      invoice.addStatusHistory(invoiceStatus, new Date(), validationVoucher.messages.join(' >> '));
-    } else if (invoice.status === InvoiceStatus.SENT) {
-      const authorizationVoucher = await this.sriAuthorizationPort.authorizeXml(invoice.accessCode);
-      const invoiceStatus =
-        authorizationVoucher.status === AuthorizationVoucherStatus.AUTHORIZED
-          ? InvoiceStatus.AUTHORIZED
-          : InvoiceStatus.REJECTED;
-      invoice.addStatusHistory(
-        invoiceStatus,
-        new Date(),
-        authorizationVoucher.messages.join(' >> '),
-      );
-    }
-    await this.invoiceRepository.updateInvoice(invoice);
-    await this.messageProducer.sendMessage(
-      new InvoiceMessage(invoice.id, invoice.orderId, invoice.status),
-    );
-  }
-}
