--- src/modules/domain/constants.ts
@@ -1,17 +1,0 @@
-export enum InvoiceStatus {
-  CREATED = 'created',
-  SIGNED = 'signed',
-  SENT = 'sent',
-  AUTHORIZED = 'authorized',
-  REJECTED = 'rejected',
-}
-
-export enum ValidationVoucherStatus {
-  ACCEPTED = 'ACCEPTED',
-  REJECTED = 'REJECTED',
-}
-
-export enum AuthorizationVoucherStatus {
-  AUTHORIZED = 'AUTHORIZED',
-  REJECTED = 'REJECTED',
-}
