--- src/modules/domain/models/index.ts
@@ -1,4 +1,0 @@
-export * from './order';
-export * from './message';
-export * from './invoice';
-export * from './voucher';
