--- src/modules/domain/ports/index.ts
@@ -1,1 +1,0 @@
-export * from './core';
