--- src/modules/infra/adapters/index.ts
@@ -1,1 +1,0 @@
-export * from './core';
