--- src/modules/infra/validators/index.ts
@@ -1,3 +1,0 @@
-export * from './core';
-export * from './message';
-export * from './sealify';
