--- src/modules/infra/validators/sealify.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod';
-
-export const SealInvoiceResponseSchema = z.object({
-  sealedData: z.string(),
-});
-
-export type SealInvoiceResponse = z.infer<typeof SealInvoiceResponseSchema>;
