+++ src/modules/invoice/app/handlers/order-message.handler.ts
@@ -1,0 +1,9 @@
+import { CreateInvoiceCommand, OrderMessage } from '../commands/create-invoice';
+
+export class OrderMessageHandler {
+  constructor(private createInvoiceCommand: CreateInvoiceCommand) {}
+
+  async handle(message: OrderMessage): Promise<void> {
+    await this.createInvoiceCommand.execute(message);
+  }
+}
