+++ src/modules/invoice/domain/invoice.ts
@@ -1,0 +1,34 @@
+import { v4 as uuidv4 } from 'uuid';
+
+export enum InvoiceStatus {
+  CREATED = 'created',
+  SIGNED = 'signed',
+  SENT = 'sent',
+  AUTHORIZED = 'authorized',
+  REJECTED = 'rejected',
+}
+
+export class InvoiceStatusHistory {
+  constructor(
+    public name: InvoiceStatus,
+    public statusDate: Date,
+    public description?: string,
+  ) {}
+}
+
+export class Invoice {
+  constructor(
+    public id: string = uuidv4(),
+    public orderId: string,
+    public accessCode: string,
+    public status: InvoiceStatus,
+    public signatureId: string,
+    public xml: string,
+    public statusHistory: InvoiceStatusHistory[] = [],
+  ) {}
+
+  addStatusHistory(status: InvoiceStatus, date = new Date(), description?: string) {
+    this.status = status;
+    this.statusHistory.push(new InvoiceStatusHistory(status, date, description));
+  }
+}
