+++ src/modules/invoice/domain/repository.ts
@@ -1,0 +1,7 @@
+import { Invoice } from './invoice';
+
+export interface InvoiceRepository {
+  createInvoice(invoice: Invoice): Promise<Invoice>;
+  updateInvoice(invoice: Invoice): Promise<Invoice>;
+  getInvoiceById(id: string): Promise<Invoice | null>;
+}
