+++ src/modules/invoice/domain/voucher.ts
@@ -1,0 +1,26 @@
+export enum ValidationVoucherStatus {
+  ACCEPTED = 'ACCEPTED',
+  REJECTED = 'REJECTED',
+}
+
+export enum AuthorizationVoucherStatus {
+  AUTHORIZED = 'AUTHORIZED',
+  REJECTED = 'REJECTED',
+}
+
+export class ValidationVoucher {
+  constructor(
+    public readonly status: ValidationVoucherStatus,
+    public readonly code: string | null,
+    public readonly messages: string[] = [],
+  ) {}
+}
+
+export class AuthorizationVoucher {
+  constructor(
+    public readonly status: AuthorizationVoucherStatus | null,
+    public readonly code: string | null,
+    public readonly authorizationDate: Date | null,
+    public readonly messages: string[] = [],
+  ) {}
+}
